refactor: move dates logic into the hooks
