+++ src/constants/errorMessages/errorMessages.ts
@@ -1,0 +1,5 @@
+export const errorMessages = {
+  datepickerFormat: 'Date should be in format dd/mm/yyyy',
+  rangepickerFormat: 'Date should be in format dd/mm/yyyy - dd/mm/yyyy',
+  datesValidation: 'Start date should be less than finish date',
+};
